+++ core/src/main/java/org/realityforge/braincheck/StackTraceUtil.java
@@ -1,17 +1,14 @@
 package org.realityforge.braincheck;
 
-import javax.annotation.Nonnull;
 
 final class StackTraceUtil
 {
-  @Nonnull
   private static final StackTraceProvider PROVIDER = new StackTraceProvider();
 
   private StackTraceUtil()
   {
   }
 
-  @Nonnull
   static StackTraceElement[] getStackTrace( final int frameCountToDrop )
   {
     return PROVIDER.getStackTrace( frameCountToDrop + 1 );
@@ -22,7 +19,6 @@ final class StackTraceUtil
   {
     @SuppressWarnings( "NonJREEmulationClassesInClientCode" )
     @GwtIncompatible
-    @Nonnull
     @Override
     StackTraceElement[] getStackTrace( final int frameCountToDrop )
     {
@@ -36,7 +32,6 @@ final class StackTraceUtil
 
   private static abstract class AbstractStackTraceProvider
   {
-    @Nonnull
     StackTraceElement[] getStackTrace( final int frameCountToDrop )
     {
       return new StackTraceElement[ 0 ];